requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==23.2.0
babel==2.17.0
base58==2.1.1
beautifulsoup4==4.13.4
bitarray==3.0.0
bleach==6.2.0
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
construct==2.10.68
construct-typing==0.5.6
contourpy==1.2.1
cryptography==44.0.0
cv2-extras==0.5.4
cycler==0.12.1
cytoolz==1.0.1
debugpy==1.8.14
decorator==4.0.11
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.0
eth-account==0.13.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.1.0
eth_abi==5.2.0
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
gast==0.5.4
ghp-import==2.1.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.28.1
idna==3.7
imageio==2.1.2
InstagramAPI==1.0.2
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.6
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.12.0
jsonalias==0.1.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
keras==3.3.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.1.3
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
ml-dtypes==0.3.2
moviepy==0.2.3.2
multidict==6.0.5
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.4.1
notebook_shim==0.2.4
numpy==1.26.4
OBP-logging-v1==0.0.1
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parsimonious==0.10.0
parso==0.8.4
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==4.25.3
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycares==4.5.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2025.1
pyunormalize==16.0.0
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyyaml_env_tag==0.1
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rlp==4.0.1
rpds-py==0.24.0
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
selenium==4.28.0
Send2Trash==1.8.3
setuptools==70.1.0
six==1.16.0
sniffio==1.3.1
solana==0.36.5
solders==0.25.0
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
terminado==0.18.1
textblob==0.19.0
threadpoolctl==3.5.0
tifffile==2024.6.18
tinycss2==1.4.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.11.2
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.1
undetected-chromedriver==3.5.5
uri-template==1.3.0
urllib3==2.2.2
virtualenv==20.27.0
watchdog==6.0.0
wcwidth==0.2.13
web3==7.7.0
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.14
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
